Move opinfo tests to separate folder
