+++ tests/opinfo/opinfo_core.py
