+++ tests/opinfo/opinfo_framework.py
