+++ tests/opinfo/opinfo_fusion_definitions.py
