+++ tests/opinfo/opinfo_input_generators.py
