+++ tests/opinfo/opinfos.py
